Extract MetadataLogProcessor module
