+++ lib/line-builders.js
@@ -15,15 +15,6 @@ function appendTime (lineParts, { prettified }) {
   }
 }
 
-function appendMetadata (lineParts, { prettified }) {
-  const { prettifiedMetadata } = prettified
-  if (prettifiedMetadata) {
-    lineParts.push(' ')
-    lineParts.push(prettifiedMetadata)
-    lineParts.push(':')
-  }
-}
-
 function appendSemicolon (lineParts) {
   if (lineParts.length > 0 && lineParts[lineParts.length - 1] !== ':') {
     lineParts.push(':')
@@ -63,7 +54,6 @@ function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys,
 
 const builtInLineBuilders = {
   appendTime,
-  appendMetadata,
   appendSemicolon,
   appendEOL,
   appendObjectOrError
