+++ lib/log-processors.js
@@ -1,4 +1,4 @@
-const { prettifyMetadata, prettifyTime } = require('./utils')
+const { prettifyTime } = require('./utils')
 const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
@@ -7,9 +7,9 @@ const SearchLogProcessor = require('./processors/SearchLogProcessor')
 const IgnoreLogProcessor = require('./processors/IgnoreLogProcessor')
 const LevelLogProcessor = require('./processors/LevelLogProcessor')
 const MessageLogProcessor = require('./processors/MessageLogProcessor')
+const MetadataLogProcessor = require('./processors/MetadataLogProcessor')
 
 const {
-  appendMetadata,
   appendSemicolon,
   appendTime,
   appendEOL,
@@ -28,16 +28,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class MetadataLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return prettifyMetadata(input, context)
-  }
-
-  build (lineParts, context) {
-    return appendMetadata(lineParts, context)
   }
 }
